+++ backend/packages/caching/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @cat-factory/caching
 
+## 0.4.22
+
+### Patch Changes
+
+- Updated dependencies [ecbcbec]
+  - @cat-factory/kernel@0.102.0
+
+## 0.4.21
+
+### Patch Changes
+
+- @cat-factory/kernel@0.101.2
+
 ## 0.4.20
 
 ### Patch Changes