+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @cat-factory/provider-s3
 
+## 0.2.103
+
+### Patch Changes
+
+- Updated dependencies [ecbcbec]
+  - @cat-factory/kernel@0.102.0
+
+## 0.2.102
+
+### Patch Changes
+
+- @cat-factory/kernel@0.101.2
+
 ## 0.2.101
 
 ### Patch Changes